requirements.txt
ollama
discord.py
python-dotenv 
fastapi 
uvicorn
requests